@@ -6,41 +6,41 @@
     "packages/*"
   ],
   "devDependencies": {
-    "@types/chai-as-promised": "7.1.7",
-    "@types/chai-jest-snapshot": "1.3.7",
-    "@types/istanbul": "0.4.32",
-    "@types/karma": "6.3.6",
-    "@types/mocha": "10.0.3",
-    "@types/node": "18.18.6",
-    "@types/sinon": "10.0.20",
-    "@types/sinon-chai": "3.2.11",
-    "@types/source-map-support": "0.5.9",
-    "@typescript-eslint/eslint-plugin": "6.7.5",
-    "@typescript-eslint/parser": "6.7.5",
+    "@types/chai-as-promised": "7.1.8",
+    "@types/chai-jest-snapshot": "1.3.8",
+    "@types/istanbul": "0.4.33",
+    "@types/karma": "6.3.7",
+    "@types/mocha": "10.0.4",
+    "@types/node": "20.9.0",
+    "@types/sinon": "17.0.1",
+    "@types/sinon-chai": "3.2.12",
+    "@types/source-map-support": "0.5.10",
+    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@typescript-eslint/parser": "6.10.0",
     "c8": "8.0.1",
     "chai": "4.3.10",
     "chai-as-promised": "7.1.1",
     "chai-jest-snapshot": "2.0.0",
     "concurrently": "8.2.2",
     "cross-env": "7.0.3",
-    "eslint": "8.52.0",
+    "eslint": "8.53.0",
     "eslint-config-prettier": "9.0.0",
     "eslint-import-resolver-typescript": "3.6.1",
     "eslint-plugin-import": "2.29.0",
     "eslint-plugin-prettier": "5.0.1",
-    "eslint-plugin-unicorn": "48.0.1",
+    "eslint-plugin-unicorn": "49.0.0",
     "execa": "8.0.1",
     "glob": "10.3.10",
     "install-local": "3.0.1",
     "jasmine": "5.1.0",
     "jasmine-core": "5.1.1",
     "json-schema-to-typescript": "13.1.1",
-    "lerna": "7.4.1",
+    "lerna": "7.4.2",
     "minimatch": "9.0.3",
     "mocha": "10.2.0",
     "prettier": "3.0.3",
     "rimraf": "5.0.5",
-    "sinon": "17.0.0",
+    "sinon": "17.0.1",
     "sinon-chai": "3.7.0",
     "source-map-support": "0.5.21",
     "ts-node": "10.9.1",
