@@ -1,3 +1,3 @@
-module.exports = {
-  default: '--require features/**/*.ts --publish-quiet'
-}
+export default {
+  import: ['features/**/*.ts'],
+};
