@@ -4,7 +4,7 @@ import fs from 'fs';
 import { URL } from 'url';
 
 const settings = JSON.parse(fs.readFileSync(new URL('../../stryker.parent.conf.json', import.meta.url), 'utf-8'));
-settings.dashboard.module = import.meta.url.split('/').slice(-2)[0];
+settings.dashboard.module = import.meta.url.split('/').at(-2);
 /**
  * @type {import('../api/dist/src/core/index.js').PartialStrykerOptions}
  */
