@@ -237,8 +237,8 @@ describe(ProjectReader.name, () => {
           normalizeWhitespaces(`Warning: No files found for mutation with the given glob expressions.
             As a result, a dry-run will be performed without actually modifying anything. 
             If you intended to mutate files, please check and adjust the configuration. 
-            Current glob pattern(s) used: "{src,lib}/**/!(*.+(s|S)pec|*.+(t|T)est).+(cjs|mjs|js|ts|jsx|tsx|html|vue)", 
-            "!{src,lib}/**/__tests__/**/*.+(cjs|mjs|js|ts|jsx|tsx|html|vue)".
+            Current glob pattern(s) used: "{src,lib}/**/!(*.+(s|S)pec|*.+(t|T)est).+(cjs|mjs|js|ts|jsx|tsx|html|vue|svelte)", 
+            "!{src,lib}/**/__tests__/**/*.+(cjs|mjs|js|ts|jsx|tsx|html|vue|svelte)".
             To enable file mutation, consider configuring the \`mutate\` 
             property in your configuration file or using the --mutate option via the command line.`),
         );
