@@ -1,6 +1,6 @@
 import { NodePath, types } from '@babel/core';
 
-import { JSAst, AstFormat, HtmlAst, TSAst } from '../../src/syntax/index.js';
+import { JSAst, AstFormat, HtmlAst, TSAst, SvelteAst, TemplateScript, Range } from '../../src/syntax/index.js';
 import { Mutant, Mutable } from '../../src/mutant.js';
 import { ParserOptions } from '../../src/parsers/index.js';
 import { InstrumenterOptions } from '../../src/index.js';
@@ -68,6 +68,34 @@ export function createTSAst(overrides?: Partial<TSAst>): TSAst {
     root: parseTS(rawContent, originFileName),
     ...overrides,
   };
+}
+
+export function createSvelteAst(overrides?: Partial<SvelteAst>): SvelteAst {
+  const rawContent = overrides?.rawContent ?? '<script>let name = "temp"</script><h1>hello {name}!</h1>';
+  const originFileName = overrides?.originFileName ?? 'foo.svelte';
+  return {
+    format: AstFormat.Svelte,
+    originFileName,
+    rawContent,
+    root: {
+      moduleScript: overrides?.root?.moduleScript,
+      additionalScripts: overrides?.root?.additionalScripts ?? [],
+    },
+    ...overrides,
+  };
+}
+
+export function createTemplateScript(overrides?: Partial<TemplateScript>): TemplateScript {
+  return {
+    ast: createJSAst(),
+    range: { start: 26, end: 16 },
+    isExpression: false,
+    ...overrides,
+  };
+}
+
+export function createRange(start = 0, end = 0): Range {
+  return { start, end };
 }
 
 export function createMutant(overrides?: Partial<Mutant>): Mutant {
