@@ -63,17 +63,45 @@ describe(Mutant.name, () => {
       expect(mutant.toApiMutant()).deep.include(expected);
     });
 
-    it('should offset location correctly', () => {
+    it('should offset location line correctly', () => {
       // Arrange
       const lt = findNodePath<babel.types.BinaryExpression>(parseJS('if(a < b) { console.log("hello world"); }'), (p) => p.isBinaryExpression()).node;
       const lte = types.binaryExpression('<=', lt.left, lt.right);
-      const mutant = new Mutant('1', 'bar.js', lt, { replacement: lte, mutatorName: 'barMutator' }, { position: 42, line: 4 });
+      const mutant = new Mutant('1', 'bar.js', lt, { replacement: lte, mutatorName: 'barMutator' }, { column: 0, line: 4 });
 
       // Act
       const actual = mutant.toApiMutant();
 
       // Assert
       expect(actual.location).deep.eq({ start: { line: 4, column: 3 }, end: { line: 4, column: 8 } });
+    });
+
+    it('should offset location columns when line = 0 (current line) correctly', () => {
+      // Arrange
+      const lt = findNodePath<babel.types.BinaryExpression>(parseJS('if(a < b) { console.log("hello world"); }'), (p) => p.isBinaryExpression()).node;
+      const lte = types.binaryExpression('<=', lt.left, lt.right);
+      const mutant = new Mutant('1', 'bar.js', lt, { replacement: lte, mutatorName: 'barMutator' }, { column: 42, line: 0 });
+
+      // Act
+      const actual = mutant.toApiMutant();
+
+      // Assert
+      expect(actual.location).deep.eq({ start: { line: 0, column: 45 }, end: { line: 0, column: 50 } });
+    });
+
+    it('should not offset columns when line != 0 (one of the next lines)', () => {
+      // Arrange
+      const lt = findNodePath<babel.types.BinaryExpression>(parseJS('console.log("skip line 1"); \nif(a < b) { console.log("hello world"); }'), (p) =>
+        p.isBinaryExpression(),
+      ).node;
+      const lte = types.binaryExpression('<=', lt.left, lt.right);
+      const mutant = new Mutant('1', 'bar.js', lt, { replacement: lte, mutatorName: 'barMutator' }, { column: 42, line: 0 });
+
+      // Act
+      const actual = mutant.toApiMutant();
+
+      // Assert
+      expect(actual.location).deep.eq({ start: { line: 1, column: 3 }, end: { line: 1, column: 8 } });
     });
   });
 
